concat: Avoid segment edges by cutting the middle of the overlap
In some cases the output's framerate was not the same as input's (59.97 instead of 60). Due to bug on YouTube this small difference led to horrible audio desync in long videos ((60-59.97)/60*3600 = 1.8 seconds per hour).

This problem was probably caused by lost frames on the edges of segments, however I'm still not entirely sure. This commit fixes output FPS on all of my test videos, but may cause some short artifacts at the cuts.